Add Academy ID which refers to the PK of the Person table.
The primary key of the person table is a composite key made up of the claim_id and the person_ref.
This commit adds an Academy ID attribute to the response object from the view record endpoint.

The Academy ID is a string of the form "{claim_id}-{person ref}".
This ID is deconstructed when a request is made and the two component values ('ClaimID' and 'PersonRef') are used to find the relevant record.